PFL-13 refactoring: create career component
